+++ stormpath/http/authc.py
@@ -1,5 +1,9 @@
 __author__ = 'ecrisostomo'
 
+import binascii
+import hmac
+import hashlib
+from collections import OrderedDict
 from datetime import datetime
 from urllib.parse import urlparse
 from uuid import uuid4
@@ -7,7 +11,6 @@ from stormpath.util import is_default_port, encode_url, str_query_string
 
 class Sauthc1Signer:
 
-    DEFAULT_ALGORITHM = "SHA256"
     HOST_HEADER = "Host"
     AUTHORIZATION_HEADER = "Authorization"
     STORMPATH_DATE_HEADER = "X-Stormpath-Date"
@@ -27,7 +30,7 @@ class Sauthc1Signer:
         time_stamp = time.strftime(self.TIMESTAMP_FORMAT)
         date_stamp = time.strftime(self.DATE_FORMAT)
 
-        nonce = uuid4()
+        nonce = str(uuid4())
 
         parsed_url = urlparse(request.href)
 
@@ -45,6 +48,96 @@ class Sauthc1Signer:
         method = request.http_method
         canonical_resource_path = self._canonicalize_resource_path_(parsed_url.path)
         canonical_query_string = self._canonicalize_query_string_(request)
+        canonical_headers_string = self._canonicalize_headers_(request)
+        signed_headers_string = self._get_signed_headers_(request)
+        request_payload_hash_hex = self._hash_hex_(self._get_request_payload_(request))
+
+        canonical_request = ''.join((method, self.NL,
+                                     canonical_resource_path, self.NL,
+                                     canonical_query_string, self.NL,
+                                     canonical_headers_string, self.NL,
+                                     signed_headers_string, self.NL,
+                                     request_payload_hash_hex))
+
+        id = ''.join((api_key.id, "/", date_stamp, "/", nonce, "/", self.ID_TERMINATOR))
+
+        canonical_request_hash_hex = self._hash_hex_(canonical_request)
+
+        string_to_sign = ''.join((self.ALGORITHM, self.NL,
+                                  time_stamp, self.NL,
+                                  id, self.NL,
+                                  canonical_request_hash_hex))
+
+        # SAuthc1 uses a series of derived keys, formed by hashing different pieces of data
+        k_secret = ''.join((self.AUTHENTICATION_SCHEME, api_key.secret))
+        k_date = self._sign_(date_stamp, k_secret)
+        k_nonce = self._sign_(nonce, k_date)
+        k_signing = self._sign_(self.ID_TERMINATOR, k_nonce)
+
+        signature = self._sign_(string_to_sign, k_signing)
+        signature_hex = binascii.hexlify(signature).decode()
+
+        authorization_header = ''.join((self.AUTHENTICATION_SCHEME, " ",
+                                        self._create_name_value_pair_(self.SAUTHC1_ID, id), ", ",
+                                        self._create_name_value_pair_(self.SAUTHC1_SIGNED_HEADERS, signed_headers_string), ", ",
+                                        self._create_name_value_pair_(self.SAUTHC1_SIGNATURE, signature_hex)))
+
+        request.http_headers[self.AUTHORIZATION_HEADER] = authorization_header
+
+    def _create_name_value_pair_(self, name, value):
+        return ''.join((name, '=', value))
+
+    def _sign_(self, data, key):
+
+        try:
+            byte_key = key.encode()
+        except:
+            byte_key = key
+
+        return hmac.new(byte_key, data.encode(), hashlib.sha256).digest()
+
+    def _hash_hex_(self, text):
+        return hashlib.sha256(text.encode()).hexdigest()
+
+    def _get_request_payload_(self, request):
+        return self._get_request_payload_without_query_params_(request)
+
+    def _get_request_payload_without_query_params_(self, request):
+
+        result = ''
+
+        if request.body:
+            result = request.body
+
+        return result
+
+    def _get_signed_headers_(self, request):
+
+        sorted_headers = OrderedDict(sorted(request.http_headers.items()))
+
+        result = ''
+
+        for header in sorted_headers.copy().keys():
+
+            if result:
+                result += ';' + header
+            else:
+                result += header
+
+        return result.lower()
+
+    def _canonicalize_headers_(self, request):
+
+        sorted_headers = OrderedDict(sorted(request.http_headers.items()))
+
+        result = ''
+
+        for key, value in sorted_headers.items():
+
+            result += ''.join((str(key).lower(), ':', value))
+            result += self.NL
+
+        return result
 
     def _canonicalize_query_string_(self, request):
         return str_query_string(request.query_string)
@@ -55,5 +148,3 @@ class Sauthc1Signer:
             return encode_url(resource_path)
         else:
             return '/'
-
-